컬럼에서 Long으로 저장 할 수 있게 변경

+++ src/test/java/org/herring/index/column/index/IndexTableSimpleTest.java
@@ -70,6 +70,6 @@ class IndexWriterSimpleTest {
 
         assertEquals(key, index.getKeyword());
         assertEquals(1, index.getIndexs().size());
-        assertEquals(location, (long) index.getIndexs().get(0));
+        assertEquals(location, index.getIndexs().get(0));
     }
 }
